chore: move files to ai-chat module
