+++ core/local/src/main/java/com/emotionstorage/local/di/LocalDataSourceModule.kt
@@ -1,9 +1,7 @@
 package com.emotionstorage.local.di
 
-import com.emotionstorage.data.dataSource.ChatIntroPrefsLocalDataSource
 import com.emotionstorage.data.dataSource.SessionLocalDataSource
 import com.emotionstorage.data.dataSource.UserLocalDataSource
-import com.emotionstorage.local.dataSourceImpl.ChatIntroPrefLocalDataSourceImpl
 import com.emotionstorage.local.dataSourceImpl.SessionLocalDataSourceImpl
 import com.emotionstorage.local.dataSourceImpl.UserLocalDataSourceImpl
 import dagger.Binds
@@ -24,10 +22,4 @@ abstract class LocalDataSourceModule {
     abstract fun bindSessionLocalDataSource(
         sessionLocalDataSourceImpl: SessionLocalDataSourceImpl,
     ): SessionLocalDataSource
-    // TODO 옮기기
-    @Binds
-    @Singleton
-    abstract fun bindChatIntroPrefsLocalDataSource(
-        chatIntroPrefLocalDataSourceImpl: ChatIntroPrefLocalDataSourceImpl,
-    ): ChatIntroPrefsLocalDataSource
 }
