+++ data/src/main/java/com/emotionstorage/data/di/RepositoryModule.kt
@@ -1,9 +1,7 @@
 package com.emotionstorage.data.di
 
-import com.emotionstorage.data.repoImpl.ChatIntroPrefsRepositoryImpl
 import com.emotionstorage.data.repoImpl.SessionRepositoryImpl
 import com.emotionstorage.data.repoImpl.UserRepositoryImpl
-import com.emotionstorage.domain.repo.ChatIntroPrefsRepository
 import com.emotionstorage.domain.repo.SessionRepository
 import com.emotionstorage.domain.repo.UserRepository
 import dagger.Binds
@@ -22,8 +20,4 @@ abstract class RepositoryModule {
     @Binds
     @Singleton
     abstract fun bindSessionRepository(impl: SessionRepositoryImpl): SessionRepository
-
-    @Binds
-    @Singleton
-    abstract fun bindChatIntroPrefsRepository(impl: ChatIntroPrefsRepositoryImpl): ChatIntroPrefsRepository
 }
