+++ feat/ai-chat/src/main/java/com/emotionstorage/ai_chat/data/dataSource/local/ChatIntroPrefsLocalDataSource.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.data.dataSource
+package com.emotionstorage.ai_chat.data.dataSource.local
 
 import kotlinx.coroutines.flow.Flow
