+++ feat/ai-chat/src/main/java/com/emotionstorage/ai_chat/data/dataSource/remote/ChatRemoteDataSource.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.ai_chat.data.dataSource
+package com.emotionstorage.ai_chat.data.dataSource.remote
 
 interface ChatRemoteDataSource {
     suspend fun getChatRoomId(): String
