+++ feat/ai-chat/src/main/java/com/emotionstorage/ai_chat/remote/dataSource/ChatWSDataSourceImpl.kt
@@ -1,6 +1,6 @@
 package com.emotionstorage.ai_chat.remote.dataSource
 
-import com.emotionstorage.ai_chat.data.dataSource.ChatWSDataSource
+import com.emotionstorage.ai_chat.data.dataSource.remote.ChatWSDataSource
 import com.emotionstorage.ai_chat.domain.model.ChatMessage
 import com.emotionstorage.ai_chat.remote.modelMapper.ChatMessageMapper
 import com.emotionstorage.ai_chat.remote.response.ChatMessageRequestBody
