+++ feat/ai-chat/src/main/java/com/emotionstorage/ai_chat/remote/di/RemoteDataSourceModule.kt
@@ -1,7 +1,7 @@
 package com.emotionstorage.ai_chat.remote.di
 
-import com.emotionstorage.ai_chat.data.dataSource.ChatRemoteDataSource
-import com.emotionstorage.ai_chat.data.dataSource.ChatWSDataSource
+import com.emotionstorage.ai_chat.data.dataSource.remote.ChatRemoteDataSource
+import com.emotionstorage.ai_chat.data.dataSource.remote.ChatWSDataSource
 import com.emotionstorage.ai_chat.remote.dataSource.ChatRemoteDataSourceImpl
 import com.emotionstorage.ai_chat.remote.dataSource.ChatWSDataSourceImpl
 import dagger.Binds
